refactor discovery tests
