+++ e2e-tests/IIoTPlatform-E2E-Tests/Discovery/DiscoveryTestTheory.cs
@@ -52,7 +52,7 @@ namespace IIoTPlatform_E2E_Tests.Discovery {
             TestHelper.CallRestApi(_context, Method.POST, TestConstants.APIRoutes.RegistryDiscover, body);
 
             // Validate that the endpoint can be found
-            var result = TestHelper.WaitForEndpointDiscoveryToBeCompleted(_context, _cancellationTokenSource.Token, requestedEndpointUrls: urls).GetAwaiter().GetResult();
+            var result = TestHelper.Discovery.WaitForEndpointDiscoveryToBeCompleted(_context, _cancellationTokenSource.Token, requestedEndpointUrls: urls).GetAwaiter().GetResult();
             Assert.Equal(url, ((string)result.items[0].registration.endpointUrl).TrimEnd('/'));
         }
 
@@ -74,7 +74,7 @@ namespace IIoTPlatform_E2E_Tests.Discovery {
             var reponse = TestHelper.CallRestApi(_context, Method.POST, TestConstants.APIRoutes.RegistryDiscover, body);
 
             // Validate that all endpoints are found
-            var result = TestHelper.WaitForEndpointDiscoveryToBeCompleted(_context, cts.Token, requestedEndpointUrls: urls).GetAwaiter().GetResult();
+            var result = TestHelper.Discovery.WaitForEndpointDiscoveryToBeCompleted(_context, cts.Token, requestedEndpointUrls: urls).GetAwaiter().GetResult();
             foreach (var item in result.items) {
                 Assert.Contains(((string)item.registration.endpointUrl).TrimEnd('/'), urls);
             }
@@ -89,7 +89,7 @@ namespace IIoTPlatform_E2E_Tests.Discovery {
 
             // Discover all servers
             var cts = new CancellationTokenSource(TestConstants.MaxTestTimeoutMilliseconds);
-            dynamic result = TestHelper.WaitForDiscoveryToBeCompletedAsync(_context, cts.Token, requestedEndpointUrls: endpointUrls).GetAwaiter().GetResult();
+            dynamic result = TestHelper.Discovery.WaitForDiscoveryToBeCompletedAsync(_context, cts.Token, requestedEndpointUrls: endpointUrls).GetAwaiter().GetResult();
 
             // Validate that all servers are discovered
             var applicationIds = new List<string>(endpointUrls.Count);
